qa/suites/upgrade: log_to_journald=false
In 8b95c4b7c58e0caa6ac157b36e1cbc616803da3d we set log_to_journald=false
in the cephadm config.  However, that's not present in pre-quincy builds,
which means that when we upgrade the new daemons start spamming the
teuthology.log.  Set this (with --force, since it's not valid pre-quincy)
in the config before we start the ugprade.

Signed-off-by: Sage Weil <sage@newdream.net>