qa/suites/upgrade: log_to_journald=false
In 2616ee26fd7a5d89cec8d7b0f22898098a487d2b we set log_to_journald=false
in the cephadm config.  However, that's not present in pre-quincy builds,
which means that when we upgrade the new daemons start spamming the
teuthology.log.  Set this (with --force, since it's not valid pre-quincy)
in the config before we start the ugprade.

Signed-off-by: Sage Weil <sage@newdream.net>