mgr/cephadm: replace multiprocessing.pool.ThreadPool with ThreadPoolExecutor
this change fixes multiprocessing pool deadlock.

Note: A ThreadPool shares the same interface as Pool, which is designed around a pool of processes and predates the introduction of the concurrent.futures module. As such, it inherits some operations that don’t make sense for a pool backed by threads, and it has its own type for representing the status of asynchronous jobs, AsyncResult, that is not understood by any other libraries.
Users should generally prefer to use concurrent.futures.ThreadPoolExecutor, which has a simpler interface that was designed around threads from the start, and which returns concurrent.futures.Future instances that are compatible with many other libraries, including asyncio.

Reference:
https://docs.python.org/3/library/multiprocessing.html#module-multiprocessing.dummy

Signed-off-by: jianglong01 <jianglong01@qianxin.com>