cls/rbd: make group_snap_list return snapshots in creation order
To make this possible the snapshot order is stored on a snapshot
creation and the shapshot list result is sorted using this order
before returning to the caller.

This change adds a restriction on start_after parameter: now it
may be either the empty string id (to start at the beginning) or
an existing snapshot id. Otherwise -ERESTART is returned to inform
the caller that the "start_after" snapshot is invalid/removed and
the listing should be restarted.

Fixes: https://tracker.ceph.com/issues/51686
Signed-off-by: Mykola Golub <mgolub@suse.com>
Signed-off-by: N Balachandran <nibalach@redhat.com>