rgw:auth: prototype rgw::auth::Identity::Predicate
Propose an abstract mechanism to effciently assert a predicate
on an rgw::auth::Identity, which is already somewhat specialized
and related to (ahem) the auth mech hierarchy, and available as
req_state->auth.identity.

The challenge with the approach is that presently it's challenging
to share state between parts of and across the auth process, so
the intent here is to support some adjustment there in following
commits.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>