qa: stop overriding distro for k-testing
This is a continuation of previous commit

    qa: only use RHEL for workload testing

We don't want to test fs:workload with centos/ubuntu to avoid packaging
issues and to reduce the matrix of distros we're running workloads on.
Also, the testing kernel should install fine on the distros we test with
"supported" random distros.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>
(cherry picked from commit fb75ed6d391960f0826ac810b942afd2f0a662ea)