cmake: make rgw_common a static library
it was changed from a shared library to object library back in
5ac36819541ffc15844b3ca615c40a47dd11fc5f, because we wanted to link both
rgw_a and rgw_admin_user against it.

but now that rgw_admin_user was dropped in
148839bea9d8b1bce5ae282fbbe21067854ad291, let's compile rgw_common
as a static library. it'd be much easier for us to apply various compile
options and linkages to it this way.

Signed-off-by: Kefu Chai <kchai@redhat.com>