vstart.sh: set objectstore to "cyanstore" if --cyanstore is specified
in e6ed65db8b4e0a2f8026c2e35a12dd292c5f2b8c, "cyanstore" is added to the
help message of vstart.sh, but we should also check for this option, and
set the ceph option accordingly.

in this change, the option is checked and "objectstore" is updated
accordingly.

Signed-off-by: Kefu Chai <kchai@redhat.com>