tools/ceph_objectstore_tool: pass require_rollback to write_log_and_missing()
before this change, `require_rollback` is always true when it is passed
to write_log_and_missing(), which does not hurt. but this diverges from
how `PG::prepare_write()` is implemented. and probably, more important,
this is not correct, as replica pool does not `require_rollback`.

after this change, `require_rollback` is deduced from pgid. so we don't
always pass `true` to `write_log_and_missing()`.

Signed-off-by: Kefu Chai <kchai@redhat.com>