rgw:If the endpoint of the event notification does not respond, the process will be stuck.
We use the function of uploading objects and set an event notification for the bucket. However, due to the wrong endpoint setting, the file cannot be uploaded and is stuck all the time without reporting an error.
So we added the CURLOPT_CONNECTTIMEOUT parameter to control the connection timeout.

fix: https://tracker.ceph.com/issues/51855

Signed-off-by: wangyingbin <ybwang0211@163.com>