debian/control: make cryptsetup-bin a ceph-volume runtime dep
was originally added by fc9a041d5dc04e60fe49c59842f238fce0345162 and
2118ea4c8dbcc3f90fa7a2c10580ff46bcfbedb1 as a dependency of ceph-disk.
although, ceph-disk was replaced by ceph-volume, cryptsetup-bin is still
used.

but it should be a runtime dependency of ceph-volume. and it is not a
build dependency of ceph, so it's removed from the build dependency list
as well.

Signed-off-by: Kefu Chai <kchai@redhat.com>