rgw: default ms_mon_client_mode = secure
Require a secure connection to the monitor.  This is important because
we may store SSL certs on the mon.

Note that secure mode was introduce before Nautilus, so this can be
backported at least to Pacific.

Fixes: https://tracker.ceph.com/issues/52000
Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 7e22d2a31d277ab3eecff47b0864b206a32e2332)