mgr/telemetry: improve the formatting of perf histograms in telemetry report
Currently, the JSON formatter displays lists vertically rahter than horizontally, which makes it particularly difficult for users to scroll through the perf histograms, much less read/interpret them. With this commit, I wrote a patch within the show() function (the function that is controlled by `ceph telemetry show`) which formats the the lists of ranges and values in the perf histograms as strings. This way, parts of the histogram output is displayed horizontally instead of vertically. This makes the Telemetry report much more human-readable.

Signed-off-by: Laura Flores <lflores@redhat.com>