rbd-nbd: add ability to quiesce on signal
Problem:
-------
Scenario possible most commonly with the new attach command.

Say we had mapped the image and mounted it, and now the application is
performing IO to the device, while the IO is ongoing it might be possible
that the rbd-nbd process received a detach request or a SIGTERM (this is
very common if the rbd-nbd is running inside a container, if there is a
container restart then it will result in SIGTERM). It is very much possible
that we hit a filesystem corruption because it might be possible that
some of the writes are still waiting at the filesystem or the block
layers, which have not reached rbd-nbd layer yet.

This can be easily handled:
1. If the application can send sync before killing the rbd-nbd process,
   but that cannot be guaranteed all the time, because we should be ready
   for ungraceful shutdown nature.
2. Also we might not want to afford synchronous I/O (mount -osync)

Also what about the IO which might be still continuing on the mount while
process is down? We cannot simply assume that the applications are always
aware of rbd-nbd process state.

One thing we all agree with is that we should always give highest priority
to the data consistency.

Solution:
--------
Freeze the filesystem as part of the detach (SIGTERM handling) so that:
1. All the data is flushed from filesystem till backend, so that data
   is guaranteed to be consistent across.
2. IO on the mount, while rbd-nbd process is down is blocked waiting for the
   filesystem.
3. Snapshots on the image, while rbd-nbd process is down can also be
   consistent.
4. Just in case if the rbd-nbd process cannot be reattached, no problem! we
   have a safe copy.

By the way, this is provided just as an option `--quiesce-onsignal` (read as
quiesce on signal) so that default is still the same as old.

Fixes: https://tracker.ceph.com/issues/51820
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>