rgw: Use 100-continue in OPA requests to reduce latency
When we send a POST request to Open Policy Agent, we send the headers first,
wait for TCP ACK from the server, and then send the body. But due to [delayed
ACK], the TCP ACK can be delayed by up to 500 ms (but typically 40 ms on
Linux), which in this case is huge as OPA typically takes only 1 ms to process
a request.

This commit adds the

```
Expect: 100-continue
```

header to the `POST` requests sent to OPA, which makes the server reply with

```
HTTP/1.1 100 Continue
```

immediately after receiving the headers, thereby avoiding the 40 ms delay.

[delayed ACK]: https://en.wikipedia.org/wiki/TCP_delayed_acknowledgment

Fixes: https://tracker.ceph.com/issues/52067
Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>