osd/scrub: only telling the scrubber of 'updates' events if these events are waited for
Only the Primary, and only when there is an active scrubbing going on and the
scrubber state machine is in WaitUpdates states, should be notified of 'updates'.
The extra messages we queued and processed earlier, apart from creating redundant
log lines and wasting CPU time, were increasing the chance of a race in the handling
of stale scrubs.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>
(cherry picked from commit 0aed4a5dd63b51288f1e01ca73d9e3632a1238be)