crimson/common: instantiate interrupt_cond in .cc
so we can explicitly instantiate it.

this should address the segfault when accessing interrupt_cond when
it is defined as a plain thread local storage template variable in the
header file.

it seems Clang is not able to identify the access to TLS variable and
the value of %fs segment register of the main thread is always zero if
interrupt_cond is defined as a plain global variable stored in
thread local storage.

Signed-off-by: Kefu Chai <kchai@redhat.com>