mgr/dashboard,prometheus: fix handling of server_addr
Commit d79b9b1 corrected the URI published by dashboard/prometheus
to use get_mgr_ip() if the address was the wildcard (eg ::).
However, the change also affected the cherrypy server.socket_host
config value, so the modules could no longer bind to a wildcard.

This patch corrects the commit to not affect the cherrypy config.

It also further corrects the dashboard URI fix to handle the 0.0.0.0
wildcard case (prometheus already had this behavior).

Fixes: d79b9b196c255166483823aad898e9cf175159dd
Fixes: http://tracker.ceph.com/issues/52002
Signed-off-by: Scott Shambarger <devel@shambarger.net>