cmake: set CMAKE_BUILD_TYPE only if it is not specified
this change address the regression introduced by a7e3ece459111d157a20d05de3a92cf4dab6bde6.
which sets CMAKE_BUILD_TYPE to "Debug" even if it is specified by
user. so the Debug flavor is built no matter what CMAKE_BUILD_TYPE's
value is as long as .git exists.

in this change, the CMAKE_BUILD_TYPE is set to Debug only if
CMAKE_BUILD_TYPE is not specified and .git exists.

Fixes: https://tracker.ceph.com/issues/52053
Signed-off-by: Kefu Chai <kchai@redhat.com>