mds: ensure that we send the btime in cap messages
While troubleshooting a different problem, I noticed that the kclient
would occasionally not report the btime in statx calls. It's most
evident in the case where two different clients are alternately writing
to the same file, and another task on one of them is repeatedly issuing
statx() calls against it.

The problem is that we're not correctly populating the MClientCaps btime
field when issuing caps, and that causes the MDS to incorrectly send a
zeroed-out btime field.

Fixes: https://tracker.ceph.com/issues/52123
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 9c33a590466f7c553ea6bc62e90dc67728fa1e26)