doc: rbd-nomad.rst -- ceph-csi controller update
This PR corrects misspellings and improves the
English in the front-matter level text under the
heading "Create Ceph-csi Controller and Plugin Nodes"
in the "RBD & Nomad Integration" chapter of the
RBD Guide.

There is some text in here that refers to the
compatibility of ceph-csi with what I think are
container versions, but this unsureness on my part
reveals what I think is a technical ambiguity in
the text. I expect that we will have to return
to this to correct it.

Signed-off-by: Zac Dover <zac.dover@gmail.com>