kv/rocksdb_cache: drop ROCKSDB_PRIszt
ROCKSDB_PRIszt is a macro for "zu", which is in turn use for printing an
(unsigned) size_t variable.

there is no point having it in the header file or define a macro for it,
as %zu is standard compliant, and we don't get any advantage by using a
macro for it.

Signed-off-by: Kefu Chai <kchai@redhat.com>