rgw: jumbo etag fix-em-up logic
The code that reads and writes etags in radosgw is currently not consistent
about whether it terminates with a nul.  The new standard is to never
append a nul.  There's a companion commit to this that fixes one
last case for that.

Etag entries can come from omap entries, mostly fetched via cls logic
from the osd.  Those also need to be fixed.

Etags for multipart uploads are of the form "md5checksum-partno"
In a mixed luminous / nautilus environment, these checksums too
can end up with an appended "\016" character.  This patch
contains revised etag fixing logic that recognizes multipart
upload etags and removes trailing junk.

Fixes: https://tracker.ceph.com/issues/61263

Signed-off-by: Marcus Watts <mwatts@redhat.com>