mds: forbid the same client to acquire the xlock for filelock
When doing the open truncate the mds will set the filelock's state
to LOCK_XLOCKDONE just before journaling the "open truncate" mdlog,
and then releases the mds_lock.

After that if the loner client send another request, which will
also try to acquire xlock for the filelock it will switch the lock
state to LOCK_LOCK_XLOCK. So when the "open truncate" mdlog is
flushed the callback will crash the mds daemon.

We need to suspend the filelock's state if it needs to revoke the
Fb cap to flush the snap data from clients before the inode is
truncated.

Fixes: https://tracker.ceph.com/issues/49132
Signed-off-by: Xiubo Li <xiubli@redhat.com>