Use consistent condition for ppc probing.
According to src/arch/CMakeLists.txt the ppc probing code is compiled
when HAVE_POWER8 is true. The condition for when checking the ppc
probe result in other places in the code must therfore also check
if HAVE_POWER8 is defined. Otherwise compilation fails with

/usr/bin/ld: ../lib/libceph-common.so.0: undefined reference to `ceph_arch_ppc_probe'
/usr/bin/ld: ../lib/libceph-common.so.0: undefined reference to `ceph_arch_ppc_crc32'
/usr/bin/ld: ../lib/libceph-common.so.0: undefined reference to `ceph_crc32c_ppc'
collect2: error: ld returned 1 exit status

The above error is from building on Debian for 32 bit powerpc.

Signed-off-by: Mattias Ellert <mattias.ellert@physics.uu.se>