client: choose the auth mds to avoid possible forward traffics
For some client requests if we send them to the none auth mds, then
in the mds side it may forward that requst to auth mds directly or
when if it needs to do the authpin or xlock it will send extra
requests to the auth mds, which will slow down the perf.

In client side always try to use the auth mds first if possible for
those requests.

Signed-off-by: Xiubo Li <xiubli@redhat.com>