tests/crimson: don't be so verbose when run by the 'make check' bot.
This commit is a try to balance crimson's debug facilities with overwhelming
some hosts the `make check` bot runs on with huge number of log entries;
see https://jenkins.ceph.com/job/ceph-pull-requests-arm64/17127/ for an
exemplification of the problem.

This is achieved by differentiating debug levels depending upon the
`FOR_MAKE_CHECK` env variable. It's a hack / a makeshift solution
only. IMHO ultimately we should introduce a dedicated target for
the bot while leaving `ninja test` for humans which would allow for more
debugs and maybe also for running `vstart.sh`-dependent tests.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>