mgr/dashboard: Fix failing config dashboard e2e check
Recently a new osd config has been added in
6ca32bde2e1d0dd58df168126582a570ac09aad6 and this is getting Modified.
So on our dashboard e2e config check which checks for the Modified
filter, this is also coming on the entry. So we need to increase the
count.

Fixes: https://tracker.ceph.com/issues/52649
Signed-off-by: Nizamudeen A <nia@redhat.com>
(cherry picked from commit 53f78d330c98f803c09a2314397d48f253db0866)